Extract parsing function into utils/files
