+++ src/types.ts
@@ -1,5 +1,7 @@
 import { UploadType } from 'multinet';
 
+export type CSVColumnType = 'label' | 'category' | 'number' | 'date';
+
 export interface App {
   name: string;
   url: string;
